salting and hashing passwords added
passwords are now salted and hashed using bcrpyt. Hash is stored instead of password in database and is checked against the hash of the login attempt password. initial hashing was added to the signup function in controller.js, password field was modified to 60 character limit to account of hash size in model.js. checking was added to login function in controller, needed to define a new findUser function that only returned username in order to get the password form the database separately.